Fixed long-lasting bug on collisional ionization

+++ src/Collisions/CollisionalIonization.cpp
@@ -184,7 +184,7 @@ void CollisionalIonization::prepare2(Particles *p1, int i1, Particles *p2, int i
 }
 
 // Method to prepare the ionization
-void CollisionalIonization::prepare3(double timestep, int n_patch_per_cell)
+void CollisionalIonization::prepare3(double timestep, double n_patch_per_cell)
 {
     // Calculate the coeff used later for ionization probability
     if( nei<=0. ) {
